Fixed rotation for certain angles
Only the min point and max point of the AABB were transformed. For this
reason rotation of about 90 degrees lead to the alignment of the two
points thus creating a transformed AABB that was non-existent. Now every
point is multiplied by the matrix and the transformed AABB is calculated
correctly