HUGE shapes refactor
Boxes are now AABB of length 1 that will be transformed with matrices.
Meshes have lost the area member. The random point is sampled using the
densities array and is aware of the non-uniform scaling of the
transformation matrix. However, spheres still (and probably will)
exhibit undefined behaviour when non-uniformly scaled