EWA WEIGHTS computed at compile time
EWA filtering is implemented, but still wrong. Its weights are computed
at compile time in a lookup table and for this reason the entire
project is shifted to C++14 (std::index_sequence). With this decision it
is possible to change the number of samples of the EWA lookup table or
the alpha value used without manually changing the values inside the
lookup table itself