Fixed linking
The extern "C" fucker works only if it is the first of the #include
list. I dunno why, I don't want to even know why, I just lost 3 days
figuring out why the functions were mangled with C++ names after
explicitly telling to the linker that the mangling was the C one.

Also, purged imagemagick from the realm of existence, I don't like it a
nd I don't want to support it anymore.