Changed Image write tests
img_write tests depends on img_read to check correctness. For this reason,
if the read routine is flawed, also the write tests will fail. However,
this guarantees that the written data is the expected one, does not
depend on popen and works also on win32.

Rewritten also a portion of the bmp save/write.