Added getters to Filter and Camera class
This avoid totally breaking encapsulation during tests.
There is a problem with the camera though. The perspective camera
documentation says that the fov is expected in radians. The perspective
camera reference for the parser says that the fov is expected in
degrees. However, I don't see anywhere a cast from parsed degrees to radians.
Maybe that's why the differential rays (and thus trilinear and EWA filtering)
are bugged as hell but only with perspective camera