Switched to CTest
The parser is not reentrant (pure) and after reworking the includes some
strange errors appear while parsing. By design this should NEVER EVER
happen during execution, but some tests fails because the order of the
tests actually matter (for example an ID is asserted but the ID is
global and sequentially incremented, thus order dependent). I discovered
this by running the tests shuffled only now, and it is too late to
change. However, given the fact that singletons are required for this
project and reworking the entire design is an useless effort, tests will
be run sequentially from now on.